will delete the folder each time you update the program fixes bug

+++ src/com/gilad/oved/telme/Login.java
@@ -45,11 +45,13 @@ public class Login extends Activity implements OnClickListener {
 			
 			
 			File userFile = new File (Environment.getExternalStorageDirectory().getAbsolutePath() + "/ListenApp/");
-    	    String[] children = userFile.list();
-    	    for (int i = 0; i < children.length; i++) {
-    	    	new File(userFile, children[i]).delete();
-    	    }
-		    userFile.delete();
+			if (userFile != null) {
+				String[] children = userFile.list();
+				for (int i = 0; i < children.length; i++) {
+					new File(userFile, children[i]).delete();
+				}
+				userFile.delete();
+			}
 		}
 	}
